Filtro especialistas con horarios, formato de contraseña en registro de usuarios

+++ backend/src/test/java/com/takecare/backend/user/service/UserVerificationServiceTest.java
@@ -120,6 +120,6 @@ class UserVerificationServiceTest {
         assertThatThrownBy(() -> userVerificationService.verifyUser(1, request))
                 .isInstanceOf(RuntimeException.class);
 
-        assertThat(mockUser.getAccountVerified()).isEqualTo(0);
+        assertThat(mockUser.getAccountVerified()).isEqualTo((byte) 0);
     }
 }
